Remove the need for a redundant annotations table (#93)
* start to remove sql queries from annotations table that require casting and also duplicates the entries in the anno_* by type tables

* separate server_init_db_queries() from termdb.sql which imports app.js and causes the updated tests to launch unnecessary server process
