+++ .gitignore
@@ -52,7 +52,7 @@ client/test/wpCompileTime.js
 cairo-*/*
 canvas/*
 pixman-*/*
-**/serverconfig.json
+**/serverconfig*.json
 !/server/test/serverconfig.json
 **/server.js*
 !client/rx/server.js
